adding authorization to job views

+++ backend/job/views.py
@@ -2,9 +2,10 @@ from urllib import request
 from warnings import filters
 from django.shortcuts import render
 from django.shortcuts import get_object_or_404
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
 from django.db.models import Avg, Min, Max, Count
 from .serializer import JobSerializer
@@ -42,7 +43,9 @@ def getJob(request, pk):
     return Response(serializer.data)
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data['user'] = request.user
     data = request.data
     job = Job.objects.create(**data)
     
@@ -51,8 +54,12 @@ def newJob(request):
     return Response(serializer.data)
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You do not have permission to update this job.'}, status=status.HTTP_403_FORBIDDEN)
     
     job.title = request.data['title']
     job.description = request.data['description']
@@ -72,8 +79,12 @@ def updateJob(request, pk):
     return Response(serializer.data)
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response({'message': 'You do not have permission to remove this job.'}, status=status.HTTP_403_FORBIDDEN)
     
     job.delete()
